requirements/base.txt
Django==1.6.0
bpython==0.12
Fabric==1.8.0
Pillow==2.2.1
PyPDF2==1.19
Wand==0.3.5
boto==2.17.0
tesserwrap==0.1.3
djangorestframework==2.3.9
